Move refactored code into master directory
