+++ blues/__init__.py
@@ -7,10 +7,3 @@ BLUES
 # Define global version.
 from . import version
 __version__ = version.version
-
-# Self module imports
-from . import utils
-from . import ncmc
-from . import ncmc_switching
-from . import posedart
-from . import smartdart
