+++ blues/ncmc.py
@@ -3,585 +3,330 @@ ncmc.py: Provides the Simulation class for running the NCMC simulation.
 
 Authors: Samuel C. Gill
 Contributors: Nathan M. Lim, David L. Mobley
+
+version: 0.0.3
 """
 
 from __future__ import print_function
-import sys
-from simtk.openmm.app import *
-from simtk.openmm import *
-from blues.ncmc_switching import *
-import simtk.unit as unit
-import mdtraj as md
-import math
-from alchemy import AbsoluteAlchemicalFactory, AlchemicalState
+from openmmtools import alchemy
 import numpy as np
-from openmmtools import testsystems
 
-def get_lig_residues(lig_resname, coord_file, top_file=None):
-    """
-    Helper function to get atom indices of a ligand from a coordinate
-    and/or topology file.
-    Arguments
-    ---------
-    lig_resname: str
-        resname that you want to get the atom indicies for (ex. 'LIG')
-    coord_file:  str
-        path of coordinate file (.pdb, .gro, .h5 etc)
-    top_file: str, optional, default=None
-        path of topology file. Include if the topology is not included
-        in the coord_file
-    Returns
-    -------
-    lig_atoms : list of ints
-        list of atoms in the coordinate file matching lig_resname
-    """
+from simtk import unit, openmm
+from simtk.openmm import app
 
-    if top_file == None:
-        traj = md.load(coord_file)
-    else:
-        traj = md.load(coord_file, top=top_file)
+from blues import utils
+from blues.ncmc_switching import NCMCVVAlchemicalIntegrator
 
-    lig_atoms = traj.top.select(("resname =~ %s") % lig_resname)
-    first_res = min(lig_atoms)
-    last_res = max(lig_atoms)
-    return lig_atoms
+import sys, parmed, math, copy
+import numpy as np
+import mdtraj
+from mdtraj.reporters import HDF5Reporter
 
-def quantity_is_finite(quantity):
-    """
-    Check that elements in quantity are all finite.
-    Arguments
+class Model(object):
+    """Model provides methods for calculating properties on the
+    object 'model' (i.e ligand) being perturbed in the NCMC simulation.
+
+    Current methods calculate the object's atomic masses and center of masss.
+    Calculating the object's center of mass will get the positions and total mass.
+    Ex.
+        from blues.ncmc import Model
+        ligand = Model(structure, 'LIG')
+        ligand.calculateProperties()
+
+    Attributes
     ----------
-    quantity : simtk.unit.Quantity
-        The quantity to check
-    Returns
-    -------
-    is_finite : bool
-        If quantity is finite, returns True; otherwise False.
-    """
-    if np.any( np.isnan( np.array(quantity / quantity.unit) ) ):
-        return False
-    return True
+    ligand.resname : string specifying the residue name of the ligand
+    ligand.topology : openmm.topology of ligand
+    ligand.atom_indices : list of atom indicies of the ligand.
+    ligand.masses : list of particle masses of the ligand with units.
+    ligand.totalmass : integer of the total mass of the ligand.
 
-
-class md_reporter:
-    """
-    Class to handle error reporting. Currently obselete.
+    #Dynamic attribtues that must be updated with each iteration
+    ligand.center_of_mass : np.array of calculated center of mass of the ligand
+    ligand.positions : np.array of ligands positions
     """
 
-    def __init__(self):
-        self.traj = None
-        self.shape = None
-        self.temp_traj = None
-        self.first = 0
-        self.last  = -1
+    def __init__(self, structure, resname='LIG'):
+        """Initialize the model.
 
-    def initialize(self, nc_context):
-
-        nc_stateinfo = nc_context.getState(True, False, False, False, False, True)
-        tempPos = nc_stateinfo.getPositions(asNumpy=True)
-
-        last_x, last_y = np.shape(tempPos)
-        self.shape = (np.reshape(tempPos, (1, last_x, last_y))).value_in_unit(unit.nanometers)
-
-    def add_frame(self, nc_context):
-        if self.traj == None:
-            nc_stateinfo = nc_context.getState(True, False, False, False, False, True)
-            tempPos = nc_stateinfo.getPositions(asNumpy=True)
-
-            last_x, last_y = np.shape(tempPos)
-            self.traj = (np.reshape(tempPos, (1, last_x, last_y))).value_in_unit(unit.nanometers)
-        else:
-            nc_stateinfo = nc_context.getState(True, False, False, False, False, True)
-            tempPos = nc_stateinfo.getPositions(asNumpy=True)
-
-            last_x, last_y = np.shape(tempPos)
-            temp = (np.reshape(tempPos, (1, last_x, last_y))).value_in_unit(unit.nanometers)
-            self.traj = np.concatenate((self.traj, temp))
-    def add_temp_traj(self, nc_context):
-        if self.temp_traj == None:
-            nc_stateinfo = nc_context.getState(True, False, False, False, False, True)
-            tempPos = nc_stateinfo.getPositions(asNumpy=True)
-
-            last_x, last_y = np.shape(tempPos)
-            self.temp_traj = (np.reshape(tempPos, (1, last_x, last_y))).value_in_unit(unit.nanometers)
-        else:
-            nc_stateinfo = nc_context.getState(True, False, False, False, False, True)
-            tempPos = nc_stateinfo.getPositions(asNumpy=True)
-
-            last_x, last_y = np.shape(tempPos)
-            temp = (np.reshape(tempPos, (1, last_x, last_y))).value_in_unit(unit.nanometers)
-            if True not in np.isnan(temp):
-                self.temp_traj = np.concatenate((self.temp_traj, temp))
-    def combine_traj(self):
-        if self.traj != None and self.temp_traj != None:
-            if True not in np.isnan(self.temp_traj):
-                self.traj = np.concatenate((self.traj, self.temp_traj))
-            self.temp_traj = None
-        if self.traj == None and self.temp_traj != None:
-            self.traj = self.temp_traj[:]
-
-    def save_traj(self, output_name, topology, traj=None):
-        if traj == None:
-            traj = self.traj
-        last_top = md.Topology.from_openmm(topology)
-        broken_ncmc = md.Trajectory(xyz=traj, topology=last_top)
-        try:
-            broken_ncmc.save_dcd(output_name)
-        except ValueError:
-            print('couldnt output', output_name, 'values too large')
-
-def rand_rotation_matrix():
-    """
-    Creates a uniform random rotation matrix
-    Returns
-    -------
-    matrix_out: 3x3 np.array
-        random rotation matrix
-    """
-    rand_quat = md.utils.uniform_quaternion()
-    matrix_out = md.utils.rotation_matrix_from_quaternion(rand_quat)
-    return matrix_out
-
-def zero_masses( system, firstres, lastres):
-    for index in range(firstres, lastres):
-        system.setParticleMass(index, 0*unit.daltons)
-
-def selectres(reslist, inpcrd, prmtop):
-    t = md.load(inpcrd, top=prmtop)
-    output_list = []
-    for entry in reslist:
-        output_list.extend(t.top.select(("resid %s") % (entry)))
-    return output_list
-
-def zero_allother_masses( system, indexlist):
-    num_atoms = system.getNumParticles()
-    for index in range(num_atoms):
-        if index in indexlist:
-            pass
-        else:
-            system.setParticleMass(index, 0*unit.daltons)
-
-class SimNCMC(object):
-    def __init__(self, temperature, residueList, **kwds):
+        Parameters
+        ----------
+        resname : str
+            String specifying the resiue name of the ligand.
+        structure: parmed.Structure
+            ParmEd Structure object of the model to be moved.
         """
-        Stores parameters and methods relevant to NCMC and runs NCMC simulations.
+
+        self.resname = resname
+        self.atom_indices = self.getAtomIndices(structure, self.resname)
+        self.topology = structure[self.atom_indices].topology
+        self.totalmass = 0
+        self.masses = []
+
+        self.center_of_mass = None
+        self.positions = structure[self.atom_indices].positions
+
+    def getAtomIndices(self, structure, resname):
+        """
+        Get atom indices of a ligand from ParmEd Structure.
         Arguments
         ---------
-        temperature: simtk.unit.kelvin
-            desired temperature of simulation in Kelvins
-        residueList: list of ints
-            residue numbers that specify all the ligand atoms
-        """
-        super(SimNCMC, self).__init__(**kwds)
-
-        self.total_mass = 0
-        self.mass_list = None
-        self.residueList = residueList
-        self.acceptance = 0
-        self.md_simulation = None
-        self.dummy_simulation = None
-        self.nc_context = None
-        self.nc_integrator = None
-        self._storage = None
-        self.temperature = temperature
-        kB = unit.BOLTZMANN_CONSTANT_kB * unit.AVOGADRO_CONSTANT_NA
-        kT = kB * temperature
-        beta = 1.0 / kT
-        self.beta = beta
-
-    def get_particle_masses(self, system, residueList=None, set_self=True):
-        """
-        Finds the mass of each particle given by residueList and returns
-        a list of those particle masses as well as the total mass. If
-        set_self=True, sets corresponding SimNCMC attributes as well as
-        returning them.
-        Arguments
-        ---------
-        system: simtk.openmm.system
-            Openmm system object containing the particles of interest
-        residueList: list of ints
-            particle indices to find the masses of
-        set_self: boolean
-            if true, sets self.total_mass and self.mass_list to the
-            outputs of this function
-        """
-        if residueList == None:
-            residueList = self.residueList
-        mass_list = []
-        total_mass = 0*unit.dalton
-        for index in residueList:
-            mass = system.getParticleMass(int(index))
-            total_mass = total_mass + mass
-            mass_list.append([mass])
-        total_mass = np.sum(mass_list)
-        mass_list = np.asarray(mass_list)
-        mass_list.reshape((-1,1))
-        total_mass = np.array(total_mass)
-        total_mass = np.sum(mass_list)
-        temp_list = np.zeros((len(residueList), 1))
-        for index in range(len(residueList)):
-            mass_list[index] = (np.sum(mass_list[index])).value_in_unit(unit.daltons)
-        mass_list =  mass_list*unit.daltons
-        if set_self == True:
-            self.total_mass = total_mass
-            self.mass_list = mass_list
-        return total_mass, mass_list
-
-    def zero_masses(self, system, atomList=None):
-        """
-        Zeroes the masses of specified atoms to constrain certain degrees of freedom.
-        Arguments
-        ---------
-        system: simtk.openmm.system
-            system to zero masses
-        atomList: list of ints
-            atom indicies to zero masses
-        """
-        for index in (atomList):
-            system.setParticleMass(index, 0*unit.daltons)
-
-
-    def calculate_com(self,  pos_state, total_mass=None, mass_list=None, residueList=None, rotate=False):
-        """
-        This function calculates the com of specified residues and optionally
-        rotates them around the center of mass.
-        Arguments
-        ---------
-        total_mass: simtk.unit.quantity.Quantity in units daltons
-            contains the total masses of the particles for COM calculation
-        mass_list:  nx1 np.array in units daltons,
-            contains the masses of the particles for COM calculation
-        pos_state:  nx3 np. array in units.nanometers
-            returned from state.getPositions
-        residueList: list of int,
-            list of atom indicies which you'll calculate the total com for
-        rotate: boolean
-            if True, rotates center of mass by random rotation matrix,
-            else returns center of mass coordiantes
+        resname : str
+            String specifying the resiue name of the ligand.
+        structure: parmed.Structure
+            ParmEd Structure object of the model to be moved.
         Returns
         -------
-        if rotate==True
-        rotation : nx3 np.array in units.nm
-            positions of ligand with or without random rotation (depending on rotate)
-        if rotate==False
-        com_coord: 1x3 np.array in units.nm
-            position of the center of mass coordinate
+        atom_indices : list of ints
+            list of atoms in the coordinate file matching lig_resname
         """
-        if residueList == None:
-            residueList = self.residueList
-        if mass_list == None:
-            mass_list = self.mass_list
-        if mass_list == None:
-#            self.get_particle_masses(system=self.md_simulation.system)
-            mass_list = self.mass_list
+        ###TODO: Add option for resnum to better select residue names
+        atom_indices = []
+        topology = structure.topology
+        for atom in topology.atoms():
+            if str(resname) in atom.residue.name:
+                atom_indices.append(atom.index)
+        return atom_indices
 
-        if total_mass == None:
-            total_mass = self.total_mass
+    def getMasses(self, topology):
+        """Returns a list of masses of the atoms in the model.
 
-        #choose ligand indicies
-        copy_orig = copy.deepcopy(pos_state)
-        lig_coord = np.zeros((len(residueList), 3))
-        for index, resnum in enumerate(residueList):
-            lig_coord[index] = copy_orig[resnum]
-        lig_coord = lig_coord*unit.nanometers
-        copy_coord = copy.deepcopy(lig_coord)
-        #mass corrected coordinates (to find COM)
-        mass_corrected = mass_list / total_mass * copy_coord
-        sum_coord = mass_corrected.sum(axis=0).value_in_unit(unit.nanometers)
-        com_coord = [0.0, 0.0, 0.0]*unit.nanometers
-        #units are funky, so do this step to get them to behave right
-        for index in range(3):
-            com_coord[index] = sum_coord[index]*unit.nanometers
-        if rotate ==True:
-            for index in range(3):
-                lig_coord[:,index] = lig_coord[:,index] - com_coord[index]
-            #multiply lig coordinates by rot matrix and add back COM translation from origin
-            rotation =  np.dot(lig_coord.value_in_unit(unit.nanometers), rand_rotation_matrix())*unit.nanometers
-            rotation = rotation + com_coord
-            return rotation
+        Parameters
+        ----------
+        structure: parmed.Structure
+            ParmEd Structure object of the model to be moved.
+        """
+        masses = unit.Quantity(np.zeros([int(topology.getNumAtoms()),1],np.float32), unit.dalton)
+        for idx,atom in enumerate(topology.atoms()):
+            masses[idx] = atom.element._mass
+        totalmass = masses.sum()
+        return masses, totalmass
+
+    def getCenterOfMass(self, positions, masses):
+        """Returns the calculated center of mass of the ligand as a np.array
+
+        Parameters
+        ----------
+        structure: parmed.Structure
+            ParmEd Structure object of the model to be moved.
+        masses : numpy.array
+            np.array of particle masses
+        """
+        coordinates = np.asarray(positions._value, np.float32)
+        center_of_mass = parmed.geometry.center_of_mass(coordinates, masses) * positions.unit
+        return center_of_mass
+
+    def calculateProperties(self):
+        """Function to quickly calculate available properties."""
+        self.masses, self.totalmass = self.getMasses(self.topology)
+        self.center_of_mass = self.getCenterOfMass(self.positions, self.masses)
+
+
+class MoveProposal(object):
+    """MoveProposal provides perturbation functions for the model during the NCMC
+    simulation. Current supported methods: random rotation.
+
+    Ex.
+        from blues.ncmc import MoveProposal
+        mover = MoveProposal(model, 'random_rotation', nstepsNC)
+
+        #Get the dictionary of proposed moves
+        mover.moves
+    """
+    def __init__(self, model, method, nstepsNC):
+        """Initialize the MovePropsal object that contains functions to perturb
+        the model in the NCMC simulation.
+
+        Parameters
+        ----------
+        model : blues.ncmc.Model object
+            The object to be perturbed in the NCMC simulation.
+        method : str
+            A string of the perturbation function name (i.e 'random_rotation')
+        nstepsNC : int
+            An integer value for the number of NCMC steps performed.
+        """
+        supported_methods = ['random_rotation']
+        if method not in supported_methods:
+            raise Exception("Method %s not implemented" % method)
         else:
-        #remove COM from ligand coordinates to then perform rotation
-            return com_coord            #remove COM from ligand coordinates to then perform rotation
+            self.moves = {'model': model, 'method' : None, 'step' : 0}
+            self.setMove(method, nstepsNC)
 
-    def create_alchemicalSystem(self, coord_file, top_file, residueList=None):
+    @staticmethod
+    def random_rotation(model, nc_context):
+        """Function that performs a random rotation about the center of mass of the model during the NCMC simulation.
         """
-        Helper function to create alchemical system.
-        Arguments
-        ---------
-        coord_file: str
-            Path of amber coordinate file
-        top_file: str
-            path of amber topology file
-        residueList: list of ints
-            Indices of ligand atoms for alchemical transformation. If none are
-            specified uses the residueList in self.residueList
-        Returns
-        ---------
-        alchemical_system: simtk.openmm.System
-            The alchemically transformed system
+        ##TODO check if we can remove deepcopy
+        initial_positions = nc_context.getState(getPositions=True).getPositions(asNumpy=True)
+        positions = copy.deepcopy(initial_positions)
+        #init_model = positions[model.atom_indices]
+
+        model.positions = positions[model.atom_indices]
+        model.center_of_mass = model.getCenterOfMass(model.positions, model.masses)
+        reduced_pos = model.positions - model.center_of_mass
+
+        # Define random rotational move on the ligand
+        rand_quat = mdtraj.utils.uniform_quaternion()
+        rand_rotation_matrix = mdtraj.utils.rotation_matrix_from_quaternion(rand_quat)
+        #multiply lig coordinates by rot matrix and add back COM translation from origin
+        rot_move =  np.dot(reduced_pos, rand_rotation_matrix) * positions.unit + model.center_of_mass
+
+        # Update ligand positions in nc_sim
+        for index, atomidx in enumerate(model.atom_indices):
+            positions[atomidx] = rot_move[index]
+        nc_context.setPositions(positions)
+        positions = nc_context.getState(getPositions=True).getPositions(asNumpy=True)
+        model.positions = positions[model.atom_indices]
+
+        #print('Diff', init_model - model.positions)
+        return model, nc_context
+
+    def setMove(self, method, step):
+        """Returns the dictionary that defines the perturbation methods to be
+        performed on the model object and the step number to perform it at."""
+        self.moves['method']  = getattr(MoveProposal, method)
+        self.moves['step'] = int(step) / 2 - 1
+
+class SimulationFactory(object):
+    """SimulationFactory is used to generate the 3 required OpenMM Simulation
+    objects (MD, NCMC, ALCH) required for the BLUES run.
+    Ex.
+        from blues.ncmc import SimulationFactory
+        sims = SimulationFactory(structure, model, **opt)
+        sims.createSimulationSet()
+    """
+    def __init__(self, structure, model, **opt):
+        """Requires a parmed.Structure of the entire system and the ncmc.Model
+        object being perturbed.
+
+        Options is expected to be a dict of values. Ex:
+        nIter=5, nstepsNC=50, nstepsMD=10000,
+        temperature=300, friction=1, dt=0.002,
+        nonbondedMethod='PME', nonbondedCutoff=10, constraints='HBonds',
+        trajectory_interval=1000, reporter_interval=1000, platform=None,
+        verbose=False"""
+
+        #Structure of entire system
+        self.structure = structure
+        #Atom indicies for model
+        self.atom_indices = model.atom_indices
+
+        self.system = None
+        self.alch_system = None
+        self.md = None
+        self.alch  = None
+        self.nc  = None
+
+        self.opt = opt
+    def generateAlchSystem(self, system, atom_indices):
+        """Returns the OpenMM System for alchemical perturbations.
+
+        Parameters
+        ----------
+        system : openmm.System
+            The OpenMM System object corresponding to the reference system.
+        atom_indices : list
+            Atom indicies of the model.
         """
+        factory = alchemy.AlchemicalFactory()
+        alch_region = alchemy.AlchemicalRegion(alchemical_atoms=atom_indices)
+        alch_system = factory.create_alchemical_system(system, alch_region)
+        return alch_system
 
-        if residueList == None:
-            residueList = self.residueList
-        prmtop = openmm.app.AmberPrmtopFile(top_file)
-        inpcrd = openmm.app.AmberInpcrdFile(coord_file)
-        temp_system = prmtop.createSystem(nonbondedMethod=openmm.app.PME, nonbondedCutoff=1*unit.nanometer, constraints=openmm.app.HBonds)
-        testsystem = testsystems.TestSystem
-        testsystem.system = temp_system
-        testsystem.topology = prmtop.topology
-        testsystem.positions = inpcrd.positions
-        ligand_atoms = get_lig_residues(lig_resname='LIG', coord_file=coord_file, top_file=top_file)
-        factory = AbsoluteAlchemicalFactory(testsystem.system, ligand_atoms=ligand_atoms, annihilate_sterics=True, annihilate_electrostatics=True)
-        self.alchemical_system = factory.createPerturbedSystem()
-        return self.alchemical_system
+    def generateSystem(self, structure, nonbondedMethod='PME', nonbondedCutoff=10,
+                       constraints='HBonds', **opt):
+        """Returns the OpenMM System for the reference system.
 
-    def create_normalSystem(self, coord_file, top_file):
+        Parameters
+        ----------
+        structure: parmed.Structure
+            ParmEd Structure object of the entire system to be simulated.
+        opt : optional parameters (i.e. cutoffs/constraints)
         """
-        Helper function to create normal system.
-        Arguments
-        ---------
-        coord_file: str
-            Path of amber coordinate file
-        top_file: str
-            path of amber topology file
-            Specified uses the residueList in self.residueList
-        Returns
-        ---------
-        system: simtk.openmm.System
-            The normal system
+        system = structure.createSystem(nonbondedMethod=eval("app.%s" % nonbondedMethod),
+                            nonbondedCutoff=nonbondedCutoff*unit.angstroms,
+                            constraints=None)
+        return system
+
+    def generateSimFromStruct(self, structure, system, nIter, nstepsNC, nstepsMD,
+                             temperature=300, dt=0.002, friction=1,
+                             reporter_interval=1000,
+                             ncmc=False, platform=None,
+                             verbose=False, printfile=sys.stdout,  **opt):
+        """Used to generate the OpenMM Simulation objects given a ParmEd Structure.
+
+        Parameters
+        ----------
+        structure: parmed.Structure
+            ParmEd Structure object of the entire system to be simulated.
+        system :
+        opt : optional parameters (i.e. cutoffs/constraints)
+        atom_indices : list
+            Atom indicies of the model.
         """
+        if ncmc:
+            #Defines ncmc move eqns for lambda peturbation of sterics/electrostatics
+            functions = { 'lambda_sterics' : 'min(1, (1/0.3)*abs(lambda-0.5))',
+                          'lambda_electrostatics' : 'step(0.2-lambda) - 1/0.2*lambda*step(0.2-lambda) + 1/0.2*(lambda-0.8)*step(lambda-0.8)' }
 
-        prmtop = openmm.app.AmberPrmtopFile(top_file)
-        inpcrd = openmm.app.AmberInpcrdFile(coord_file)
-        temp_system = prmtop.createSystem(nonbondedMethod=openmm.app.PME, nonbondedCutoff=1*unit.nanometer, constraints=openmm.app.HBonds)
-        testsystem = testsystems.TestSystem
-        testsystem.system = temp_system
-        testsystem.topology = prmtop.topology
-        testsystem.positions = inpcrd.positions
-        self.normalsystem = testsystem
-        return self.normalsystem
-
-    def rotationalMove(self, context=None, residueList=None):
-        """
-        Function to be used in movekey. Performs a rotation around the center of mass
-        of the ligand. Generally no arguments are specified, since this takes the
-        context and residueList attributes from the SimNCMC class.
-
-        Arguments
-        ---------
-        context: openmm.context
-            variable referring to the ncmc context. If none are
-            specified uses the context in self.nc_context
-        residueList: list of ints
-            Indices of ligand atoms for alchemical transformation. If none are
-            specified uses the residueList in self.residueList
-
-        """
-
-        if residueList == None:
-            residueList = self.residueList
-        if context == None:
-            context = self.nc_context
-        before_rot_pos = context.getState(getPositions=True).getPositions(asNumpy=True)
-        rot_output = self.calculate_com(total_mass=self.total_mass, mass_list=self.mass_list, pos_state=before_rot_pos, residueList=residueList, rotate=True)
-        rot_output = rot_output[:].value_in_unit(unit.nanometers)
-        rotPos = before_rot_pos.value_in_unit(unit.nanometers)
-        for index, resnum in enumerate(residueList):
-            rotPos[resnum] = rot_output[index]
-        rotPos[:] = rotPos*unit.nanometers
-        context.setPositions(rotPos)
-
-    def runSim(self, md_simulation, nc_context, nc_integrator, dummy_simulation, movekey=None, nstepsNC=25, nstepsMD=1000, niter=10,
-                periodic=True, verbose=False, print_output=sys.stdout, residueList=None, alchemical_correction=False,
-                ncmc_storage='out_ncmc.h5', write_ncmc_interval=None):
-        """
-        Runs a ncmc+MD simulation
-        Arguments
-        ---------
-        md_simulation: simtk.openmm.simulation
-            normal interacting simulation
-        nc_context: simtk.openmm.context
-            Context containing the alchemical system with a NCMCAlchemicalIntegrator (currently
-            either GHMC or VV based)
-        nc_intgrator: blues.ncmc_switching.NCMCAlchemicalIntegrator
-            integrator used for the NCMC steps
-        dummy_simulation: simtk.openmm.simulation
-            A copy of the md_simulation. Is used for alchemical corrections and for writing ncmc frames
-            if write_ncmc_interval is specified
-        movekey: list of 2 item lists, the first item is a reference to a function and the second is a list of ints
-            This will perform the given function at the start of a NCMC step, given that step is listed in
-            the list of ints. In the case where mulitple moves are being performed at the same step they are performed
-            in list order.
-        nstepsNC: int, optional, default=25
-            Number of NC steps to be performed during an iteration
-        nstepsMD: int, optional, default=1000
-            Number of MD steps to be performed during an iteration
-        niter: int, optional, default=10
-            Number of total iterations to be performed
-        verbose: boolean, default=False
-            print more detailed information on each step
-        print_output: sys.stdout or str, default=sys.stdout:
-            If str outputs print statemetents to file matching str.
-            Otherwise outputs to sys.stdout
-        alchemical_correction: boolean, optional, default=False
-            Whether or not to perform alchemical correction
-        ncmc_storage: str, optional, default='out_ncmc.h5'
-            Name of hdf5 file to store positions of system during NCMC step
-            if write_ncmc_interval is specified.
-        write_ncmc_interval: int or None, optional, default=None
-            If int is used, specifies the interval which NCMC positions are written.
-            Also writes out after every function applied by movekey
-        """
-        if residueList == None:
-            residueList = self.residueList
-        self.md_simulation = md_simulation
-        self.dummy_simulation = dummy_simulation
-        self.nc_context = nc_context
-        self.nc_integrator = nc_integrator
-        self.get_particle_masses(self.md_simulation.system)
-        if print_output == sys.stdout:
-            print_file = sys.stdout
+            integrator = NCMCVVAlchemicalIntegrator(temperature*unit.kelvin,
+                                                    system,
+                                                    functions,
+                                                    nsteps=nstepsNC,
+                                                    direction='insert',
+                                                    timestep=0.001*unit.picoseconds,
+                                                    steps_per_propagation=1)
         else:
-            print_file = open(print_output, 'wb')
-        if type(write_ncmc_interval) == int:
-            h5reporter = md.reporters.HDF5Reporter(file=ncmc_storage, reportInterval=100,
-                coordinates=True, time=False, cell=False, potentialEnergy=False, kineticEnergy=False,
-                temperature=False, velocities=False, atomSubset=None)
-        #set up initial counters/ inputs
-        accCounter = 0
-        #set inital conditions
-        md_stateinfo = md_simulation.context.getState(True, True, False, True, True, periodic)
-        oldPos = md_stateinfo.getPositions(asNumpy=True)
-        oldVel = md_stateinfo.getVelocities(asNumpy=True)
+            integrator = openmm.LangevinIntegrator(temperature*unit.kelvin,
+                                                   friction/unit.picosecond,
+                                                   dt*unit.picoseconds)
+        #TODO SIMPLIFY TO 1 LINE.
+        #Specifying platform properties here used for local development.
+        if platform is None:
+            #Use the fastest available platform
+            simulation = app.Simulation(structure.topology, system, integrator)
+        else:
+            platform = openmm.Platform.getPlatformByName(platform)
+            prop = dict(DeviceIndex='2') # For local testing with multi-GPU Mac.
+            simulation = app.Simulation(structure.topology, system, integrator, platform, prop)
 
-        oldPE =  md_stateinfo.getPotentialEnergy()
-        oldKE =  md_stateinfo.getKineticEnergy()
-        nc_context.setPositions(oldPos)
-        nc_context.setVelocities(oldVel)
-        nc_stateinfo = nc_context.getState(True, False, False, False, False, periodic)
+        if verbose:
+            # OpenMM platform information
+            mmver = openmm.version.version
+            mmplat = simulation.context.getPlatform()
+            print('OpenMM({}) simulation generated for {} platform'.format(mmver, mmplat.getName()), file=printfile)
 
-        for stepsdone in range(niter):
-            print('performing ncmc step', file=print_file)
-            print('accCounter =', accCounter, file=print_file)
-            mdinfo = md_simulation.context.getState(True, True, False, True, True, periodic)
-            oldPE =  mdinfo.getPotentialEnergy()
-            oldKE =  mdinfo.getKineticEnergy()
-            if alchemical_correction == True:
-                alc_oldPE = nc_context.getState(True, True, False, True, True, periodic).getPotentialEnergy()
+            # Host information
+            #from platform import uname
+            #for k,v in uname()._asdict().items():
+            #    print(k, ':', v, file=printfile)
 
-            for stepscarried in range(nstepsNC):
-                if movekey != None:
-                    for func in movekey:
-                        if stepscarried in func[1]:
-                            print('doing the move', file=print_file)
-                            print('before step', nc_integrator.getGlobalVariableByName('lambda'), file=print_file)
+            # Platform properties
+            for prop in mmplat.getPropertyNames():
+                val = mmplat.getPropertyValue(simulation.context, prop)
+                print(prop, ':', val, file=printfile)
 
-                            func[0]()
-                            if write_ncmc_interval:
-                                positions = nc_context.getState(getPositions=True).getPositions(asNumpy=True)
-                                dummy_simulation.context.setPositions(positions)
-                                h5reporter.report(dummy_simulation, dummy_simulation.context.getState(True, True))
+        # Set initial positions/velocities
+        # Will get overwritten from saved State.
+        simulation.context.setPositions(structure.positions)
+        simulation.context.setVelocitiesToTemperature(temperature*unit.kelvin)
 
-                try:
-                    if verbose:
-                        print('work_before', nc_integrator.getGlobalVariableByName("total_work"), file=print_file)
-                        print('lambda_before', nc_integrator.getGlobalVariableByName("lambda"), file=print_file)
-                        print('shadow_before', nc_integrator.getGlobalVariableByName("shadow_work"), file=print_file)
-                        print('protocol_before', nc_integrator.getGlobalVariableByName("protocol_work"), file=print_file)
-                        print('Epert_before', nc_integrator.getGlobalVariableByName("Epert"), file=print_file)
+        ###TODO MOVE SIMULATION REPORTERS TO OWN FUNCTION.
+        simulation.reporters.append(app.StateDataReporter(sys.stdout, separator="\t",
+                                    reportInterval=reporter_interval,
+                                    step=True, totalSteps=nIter*nstepsMD,
+                                    time=True, speed=True, progress=True,
+                                    elapsedTime=True, remainingTime=True))
 
-                    if write_ncmc_interval and (stepscarried+1) % write_ncmc_interval == 0:
-                        if verbose:
-                            print('writing coordinates')
-                        positions = nc_context.getState(getPositions=True).getPositions(asNumpy=True)
-                        dummy_simulation.context.setPositions(positions)
-                        h5reporter.report(dummy_simulation, dummy_simulation.context.getState(True, True))
+        return simulation
 
-                    nc_integrator.step(1)
-                    if verbose:
-                        print('work_after', nc_integrator.getGlobalVariableByName("total_work"), file=print_file)
-                        print('lambda_after', nc_integrator.getGlobalVariableByName("lambda"), file=print_file)
-                        print('shadow_after', nc_integrator.getGlobalVariableByName("shadow_work"), file=print_file)
-                        print('protocol_after', nc_integrator.getGlobalVariableByName("protocol_work"), file=print_file)
-                        print('Eold', nc_integrator.getGlobalVariableByName("Eold"), file=print_file)
-                        print('Enew', nc_integrator.getGlobalVariableByName("Enew"), file=print_file)
-                        print('Epert_after', nc_integrator.getGlobalVariableByName("Epert"), file=print_file)
+    def createSimulationSet(self):
+        """Function used to generate the 3 OpenMM Simulation objects."""
+        self.system = self.generateSystem(self.structure, **self.opt)
+        self.alch_system = self.generateAlchSystem(self.system, self.atom_indices)
 
-                except Exception as e:
-                    if str(e) == "Particle coordinate is nan":
-                        print('nan, breaking', file=print_file)
-                        break
-
-            log_ncmc = nc_integrator.getLogAcceptanceProbability(nc_context)
-            newinfo = nc_context.getState(True, True, False, True, True, periodic)
-            newPos = newinfo.getPositions(asNumpy=True)
-            newVel = newinfo.getVelocities(asNumpy=True)
-            randnum =  math.log(np.random.random())
-            if alchemical_correction == True and np.isnan(log_ncmc) == False:
-                alc_newPE = newinfo.getPotentialEnergy()
-                dummy_simulation.context.setPositions(newPos)
-                dummy_info = dummy_simulation.context.getState(True, True, False, True, True, periodic)
-                norm_newPE = dummy_info.getPotentialEnergy()
-                correction_factor = -1.0*((norm_newPE - alc_newPE) - (oldPE - alc_oldPE))*(1/nc_integrator.kT)
-                print('correction_factor', correction_factor, file=print_file)
-                log_ncmc = log_ncmc + correction_factor
-
-            if log_ncmc > randnum:
-                print('ncmc move accepted', file=print_file)
-                if verbose:
-                    print('ncmc PE', newinfo.getPotentialEnergy(), 'old PE', oldPE, file=print_file)
-                    print('ncmc Total energy', newinfo.getPotentialEnergy() + newinfo.getKineticEnergy(), file=print_file)
-                    PE_diff = newinfo.getPotentialEnergy() - oldPE
-                    print('PE_diff', PE_diff, file=print_file)
-                print('accepted since', log_ncmc, '>', randnum, file=print_file)
-                print('log_ncmc > randnum', file=print_file)
-                print('move accepted', file=print_file)
-                accCounter = accCounter + 1.0
-                print('accCounter', float(accCounter)/float(stepsdone+1), accCounter, file=print_file)
-                nc_stateinfo = nc_context.getState(True, True, False, False, False, periodic)
-                oldPos = newPos[:]
-                oldVel = newVel[:]
-
-            else:
-                print('ncmc PE', newinfo.getPotentialEnergy(), 'old PE', oldPE, file=print_file)
-                print('rejected', log_ncmc, '<', randnum, file=print_file)
-                print('log_ncmc > randnum', file=print_file)
-                print('move rejected', file=print_file)
-                nc_context.setPositions(oldPos)
-                nc_context.setVelocities(-oldVel)
-
-            print('accCounter:', accCounter,  'iter:', stepsdone+1, file=print_file)
-            nc_integrator.reset()
-            md_simulation.context.setPositions(oldPos)
-            md_simulation.context.setVelocities(oldVel)
-            md_simulation.context.setVelocitiesToTemperature(self.temperature)
-
-            if nstepsMD > 0:
-                try:
-                    md_simulation.step(nstepsMD)
-                except Exception as e:
-                    print('Error:', e)
-                    stateinfo = md_simulation.context.getState(True, True, False, False, False, periodic)
-                    last_x, last_y = np.shape(oldPos)
-                    reshape = (np.reshape(oldPos, (1, last_x, last_y))).value_in_unit(unit.nanometers)
-                    print('potential energy before NCMC', oldPE, file=print_file)
-                    print('kinetic energy before NCMC', oldKE, file=print_file)
-
-                    last_top = md.Topology.from_openmm(md_simulation.topology)
-                    broken_frame = md.Trajectory(xyz=reshape, topology=last_top)
-                    broken_frame.save_pdb('broken.pdb')
-                    exit()
-
-            md_stateinfo = md_simulation.context.getState(True, True, False, False, False, periodic)
-            oldPos = md_stateinfo.getPositions(asNumpy=True)
-            oldVel = md_stateinfo.getVelocities(asNumpy=True)
-            nc_integrator.reset()
-            nc_context.setPositions(oldPos)
-            nc_context.setVelocities(oldVel)
-
-        acceptRatio = accCounter/float(niter)
-        print('acceptance ratio', acceptRatio, file=print_file)
-        print('numsteps ', nstepsNC, file=print_file)
-        if print_output != sys.stdout:
-            print_file.close()
-        return oldPos
+        self.md = self.generateSimFromStruct(self.structure, self.system, **self.opt)
+        self.alch = self.generateSimFromStruct(self.structure, self.system,  **self.opt)
+        self.nc = self.generateSimFromStruct(self.structure, self.alch_system,
+                                            ncmc=True, **self.opt)
