+++ blues/version.py
@@ -1,7 +1,7 @@
 
 # This file is automatically generated by setup.py
-short_version = '0.0.1'
-version = '0.0.1.dev2'
-full_version = '0.0.1.dev2-0663c58'
-git_revision = '0663c5880b17e5f55de96dc6e45347c58a599a9d'
+short_version = '0.0.3'
+version = '0.0.3.dev1'
+full_version = '0.0.3.dev1-7a17d82'
+git_revision = '7a17d82cb02a8e061a774ff1e35941a2c6120ae1'
 release = False
