--- blues_refactor/__init__.py
@@ -1,9 +1,0 @@
-#!/usr/local/bin/env python
-
-"""
-BLUES
-"""
-
-# Define global version.
-from . import version
-__version__ = version.version
