--- blues_refactor/version.py
@@ -1,7 +1,0 @@
-
-# This file is automatically generated by setup.py
-short_version = '0.0.1'
-version = '0.0.1.dev2'
-full_version = '0.0.1.dev2-04cb986'
-git_revision = '04cb98623445c8aa847b62473263632ac3fcf4e7'
-release = False
